transaction editing finished, but probably needs some clean up

+++ src/qmlapp/viewmodel/AmountVm.h
@@ -5,5 +5,5 @@
 namespace AmountVM {
 
 QString formatAmount(const int64_t amount);
-int8_t amountFromString(const QString& string);
+int64_t amountFromString(const QString& string);
 }
